doc: automated a record of all-time contributors
We want to produce a simple acknowledgement report, nothing too fancy.

Also: refactored CONTRIBUTING guidelines to refer to DCO terms.

File CONTRIBUTORS.md is updated once a week (or manually triggered).
This posts a pull request, which not auto-merged yet (there a few github
action hurdles to overcome to achieve that).

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>